Se agrega creacion de usuarios

+++ src/main/java/com/io/usyc/Repository/AppRoleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AppRoleRepository extends JpaRepository<AppRole, Integer> {
+public interface AppRoleRepository extends JpaRepository<AppRole, Long> {
     Optional<AppRole> findByCode(String code);
 }
